refactor(sidebar): refactoriza sidebar para evitar errores de compilacion
